pyproject.toml
[project]
name = "creative-automation-pipeline"
version = "0.1.0"
description = "Forward-deployed POC: a brief-to-asset creative automation pipeline. Pure pipeline core, FastAPI transport, Vue thin UI."
requires-python = ">=3.11"
authors = [{ name = "Alex Beattie" }]
license = { text = "MIT" }

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "pillow>=10.4.0",
    "openai>=1.54.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.12",
    "aiofiles>=24.1.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
]

[project.scripts]
pipeline-cli = "pipeline.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["pipeline*", "app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "ASYNC"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = [
    'ignore:Field name "copy" in "Asset" shadows an attribute in parent "BaseModel":UserWarning',
]
